LPS-189313 Create a new method to ensure that the model listener avoids indexing any data until it verifies successful indices creation.
This scenario can occur when workflow metrics indexes are being created along with other workflow definitions during the portalInstanceRegistered phase.

It is crucial to exercise caution during portal initialization and avoid relying on service.ranking in this specific situation, so WorkflowMetricsReindexPortalInstanceLifecycleListener will make sure all data will be indexed properly.

+++ modules/dxp/apps/portal-workflow/portal-workflow-kaleo-metrics-integration/src/main/java/com/liferay/portal/workflow/kaleo/metrics/integration/internal/model/listener/KaleoTaskModelListener.java
@@ -28,7 +28,10 @@ public class KaleoTaskModelListener extends BaseKaleoModelListener<KaleoTask> {
 		KaleoDefinitionVersion kaleoDefinitionVersion =
 			getKaleoDefinitionVersion(kaleoTask.getKaleoDefinitionVersionId());
 
-		if (Objects.isNull(kaleoDefinitionVersion)) {
+		if (Objects.isNull(kaleoDefinitionVersion) ||
+			!_indexerHelper.hasWorkflowMetricsIndices(
+				kaleoTask.getCompanyId())) {
+
 			return;
 		}
 
@@ -39,6 +42,12 @@ public class KaleoTaskModelListener extends BaseKaleoModelListener<KaleoTask> {
 
 	@Override
 	public void onAfterRemove(KaleoTask kaleoTask) {
+		if (!_indexerHelper.hasWorkflowMetricsIndices(
+				kaleoTask.getCompanyId())) {
+
+			return;
+		}
+
 		DeleteNodeRequest.Builder builder = new DeleteNodeRequest.Builder();
 
 		_nodeWorkflowMetricsIndexer.deleteNode(
